made sort/filter component
