+++ app/search/SearchCards.tsx
@@ -19,13 +19,17 @@ import StopIcon from '@mui/icons-material/Stop';
 import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+import Hidden from '@mui/material/Hidden';
+import { FilterListTwoTone } from '@mui/icons-material';
 import Slider from '@mui/material/Slider';
 import { useRouter } from 'next/navigation';
 import { useSearchParams } from 'next/navigation'
 import Link from 'next/link';
 import GetSpotifySearch from '../server_components/GetSpotifySearch';
 import CircleOfFifths from '../components/CircleOfFifths';
+import SortFilter from '../components/SortFilter';
 import { reverseKeyConvert } from '@/app/utils';
+
 
 const SmallPlayButton = styled(IconButton)(() => ({
 	'&&': {
@@ -115,11 +119,6 @@ const SearchCards = ({ results }) => {
 	//filter hooks
 	const [activeSlice, setActiveSlice] = useState<string | null>(null);
 	const [tempoSelect, setTempoSelect] = React.useState([0, 200]);
-	const [sliderValue, setSliderValue] = useState([80, 140]);
-	const [textFieldTempo, setTextFieldTempo] = useState('');
-	const [openAccordion, setOpenAccordion] = useState<string | null>(null);
-	const keyAccordionRef = useRef(null);
-	const bpmAccordionRef = useRef(null);
 
 	const playAudio = (event: React.MouseEvent, previewUrl: string | null) => {
 		event.stopPropagation();
@@ -167,70 +166,6 @@ const SearchCards = ({ results }) => {
 		}
 	}, [searchQuery]);
 
-	//filter effect
-	useEffect(() => {
-		setActiveSlice(null);
-		setTempoSelect([0, 200]);
-	}, [offset, searchQuery]);
-
-	//close accordion
-	useEffect(() => {
-		document.addEventListener('mousedown', handleOutsideClick);
-
-		return () => {
-			document.removeEventListener('mousedown', handleOutsideClick);
-		};
-	}, [openAccordion]);
-
-	//sorting
-	const handleSort = (attribute: "tempo" | "key") => {
-		if (sortBy === attribute && sortOrder === "asc") {
-			setSortOrder("desc");
-		} else {
-			setSortBy(attribute);
-			setSortOrder("asc");
-		}
-	};
-
-	//tempo filter
-	const valuetext = (value) => {
-		return `${value} bpm`;
-	}
-	const handleTempoSelect = (event, tempo) => {
-		setSliderValue(tempo);
-	};
-
-	const handleTempoSubmit = (event) => {
-		event.preventDefault();
-
-		if (textFieldTempo) {
-			setTempoSelect([parseFloat(textFieldTempo), parseFloat(textFieldTempo)]);
-		} else {
-			setTempoSelect(sliderValue);
-		}
-	};
-
-	const handleTextFieldChange = (event) => {
-		const value = event.target.value.replace(/[^0-9]/g, '');  // Only allow digits
-		setTextFieldTempo(value);
-	};
-	//reset filter
-	const handleReset = (event) => {
-		event.preventDefault();  // To prevent the default behavior
-		setTempoSelect([0, 200]);
-		setSliderValue([80, 140]);
-		setTextFieldTempo('');  // Clear the textfield
-		setActiveSlice('');
-	};
-	const handleOutsideClick = (event) => {
-		if (keyAccordionRef.current && !keyAccordionRef.current.contains(event.target) && openAccordion === 'keyAccordion') {
-			setOpenAccordion(null);
-		}
-
-		if (bpmAccordionRef.current && !bpmAccordionRef.current.contains(event.target) && openAccordion === 'bpmAccordion') {
-			setOpenAccordion(null);
-		}
-	};
 	return (
 		<Box>
 			<Grid container item xs={12} justifyContent='center' alignItems='center' >
@@ -270,148 +205,21 @@ const SearchCards = ({ results }) => {
 							</Card>
 						</Grid>
 
-						<Box border={1} borderColor="grey.500" borderRadius={2} m={0} sx={{
-							width: '65vw',
-							'@media (max-width: 900px)': {
-								width: '90vw',
-							}
-						}}>
-							{/* sort */}
-							<Grid item container justifyContent='center' xs={12} md={8} spacing={1}>
-								<Grid item xs={10} container alignItems="center" spacing={1}> {/* Added container and alignItems */}
-									<Grid item> {/* Wrap Typography in a Grid item */}
-										<Typography fontSize='1rem' color='white'>Sort by: </Typography>
-									</Grid>
-									<Grid item > {/* Wrap the SortButton in a Grid item */}
-										<SortButton onClick={() => handleSort("key")}>
-											<Typography fontSize='1rem' sx={{ textTransform: 'none', }}>
-												Key
-											</Typography>
-											{sortBy === "key" && sortOrder === "asc" ? "↑" : "↓"}
-										</SortButton>
-									</Grid>
-									<Grid item> {/* Wrap the next SortButton in a Grid item */}
-										<SortButton onClick={() => handleSort("tempo")}>
-											<Typography fontSize='1rem'>BPM</Typography>
-											{sortBy === "tempo" && sortOrder === "asc" ? "↑" : "↓"}
-										</SortButton>
-									</Grid>
-								</Grid>
-							</Grid>
+						<SortFilter
+							setActiveSlice={setActiveSlice}
+							activeSlice={activeSlice}
+							tempoSelect={tempoSelect}
+							setTempoSelect={setTempoSelect}
+							offset={offset}
+							searchQuery={searchQuery}
+							setSortOrder={setSortOrder}
+							sortOrder={sortOrder}
+							setSortBy={setSortBy}
+							sortBy={sortBy}
+						>
 
-							{/*/ filter */}
-							<Grid item container className='py-1' justifyContent='center' xs={12} md={8} spacing={1}>
-								<Grid item xs={10} container alignItems="center" spacing={1}> {/* Added container and alignItems */}
-									<Grid item >
-										<Typography fontSize='1rem' color='white'>Filter by:</Typography>
-									</Grid>
-									<Grid item >
-										<Accordion
-											ref={keyAccordionRef}
-											expanded={openAccordion === 'keyAccordion'}
-											onChange={() => setOpenAccordion(prev => prev === 'keyAccordion' ? null : 'keyAccordion')}
-										>
-											<StyledAccordionSummary
-												expandIcon={<ExpandMoreIcon />}
-											>
-												<Typography fontSize='0.8rem' >Key</Typography>
-											</StyledAccordionSummary>
+						</SortFilter>
 
-											<KeyAccordionDetails>
-												<Box>
-													<CircleOfFifths activeSlice={activeSlice} setActiveSlice={setActiveSlice} />
-												</Box>
-											</KeyAccordionDetails>
-
-										</Accordion>
-									</Grid>
-
-									<Grid item >
-										<Accordion
-											ref={bpmAccordionRef}
-											expanded={openAccordion === 'bpmAccordion'}
-											onChange={() => setOpenAccordion(prev => prev === 'bpmAccordion' ? null : 'bpmAccordion')}
-										>
-											<StyledAccordionSummary
-												expandIcon={<ExpandMoreIcon />}
-											>
-												<Typography fontSize='0.8rem' >BPM</Typography>
-											</StyledAccordionSummary>
-											<TempoAccordionDetails>
-												<form onSubmit={handleTempoSubmit}>
-													<Box sx={{
-														display: 'flex',
-														height: '200px',
-														width: '270px',
-														flexDirection: 'column',
-														justifyContent: 'center',
-													}}>
-														<Slider
-															min={0}
-															max={200}
-															getAriaLabel={() => 'Tempo'}
-															value={sliderValue}
-															onChange={handleTempoSelect}
-															valueLabelDisplay="on"
-															getAriaValueText={valuetext}
-														/>
-														<TextField
-															value={textFieldTempo}
-															onChange={handleTextFieldChange}
-															id="filled-basic"
-															label="select a range or enter a bpm"
-															variant="filled"
-															autoComplete="off"
-															InputProps={{
-																style: {
-																	backgroundColor: '#eceff1',
-																}
-															}}
-														/>
-														<Button type="submit" variant="contained"
-															sx={{
-																'&&': {
-																	color: 'white',
-																	backgroundColor: '#4d97f8',
-																	'&:hover': {
-																		backgroundColor: '#3746a2',
-																	},
-																}
-
-															}}
-														>
-															Filter Tempo
-														</Button>
-													</Box>
-												</form>
-
-											</TempoAccordionDetails>
-										</Accordion>
-									</Grid>
-
-									<Grid item paddingLeft={'2px'}>
-										<Box
-											onClick={handleReset}
-											variant="contained"
-											color="#ffecb3"
-											sx={{
-												// backgroundColor:'purple',
-												display: 'flex',
-												justifyContent: 'center',
-												alignItems: 'center',
-												cursor: 'pointer',
-												width: '60px',
-												height: '24px',
-
-											}}
-										>
-											Reset
-										</Box>
-
-									</Grid>
-								</Grid>
-							</Grid>
-						</Box>
 						{/* main search */}
 						{searchResults
 							.filter(item => (!activeSlice || item.key === activeSlice) && item.tempo >= tempoSelect[0] && item.tempo <= tempoSelect[1])
