+++ app/types/cardTypes.ts
@@ -1,7 +1,13 @@
-import { SongDetails } from "./dataTypes";
+import { SongDetails, AlbumDetails } from "./dataTypes";
+
 export interface SongPageCardProps {
     songDetails: SongDetails;
     song: string;
     artist: string;
     id: string;
+}
+
+export interface AlbumPageCardProps{
+    results: AlbumDetails;
+    album: string;
 }
